# vulkan: Use fp16 for the flash attention P*V multiplication