@0cc4m 2025-04-09
I don't see much of a difference on 3090, but no regression either:

Master:

| model                          |       size |     params | backend    | ngl | fa |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | -: | ------------: | -------------------: |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |  1 |        pp4096 |       3397.53 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |  1 |        pp8192 |       3084.63 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |  1 |       pp16384 |       2503.77 ± 0.00 |

```
  FLASH_ATTN_EXT(hsk=64,hsv=64,nh=8,nr=4,kv=4096,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):                 23848 runs -    45.95 us/run -  33.55 MFLOP/run - 730.26 GFLOPS
  FLASH_ATTN_EXT(hsk=128,hsv=128,nh=8,nr=4,kv=4096,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):               17892 runs -    57.68 us/run -  67.11 MFLOP/run -   1.16 TFLOPS
  FLASH_ATTN_EXT(hsk=64,hsv=64,nh=8,nr=4,kv=8192,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):                 17892 runs -    57.70 us/run -  67.11 MFLOP/run -   1.16 TFLOPS
  FLASH_ATTN_EXT(hsk=128,hsv=128,nh=8,nr=4,kv=8192,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):               12682 runs -    80.59 us/run - 134.22 MFLOP/run -   1.67 TFLOPS
  FLASH_ATTN_EXT(hsk=64,hsv=64,nh=8,nr=4,kv=16384,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):                12682 runs -    79.57 us/run - 134.22 MFLOP/run -   1.69 TFLOPS
  FLASH_ATTN_EXT(hsk=128,hsv=128,nh=8,nr=4,kv=16384,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):               8579 runs -   116.59 us/run - 268.44 MFLOP/run -   2.30 TFLOPS
```

PR:

| model                          |       size |     params | backend    | ngl | fa |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | -: | ------------: | -------------------: |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |  1 |        pp4096 |       3381.02 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |  1 |        pp8192 |       3040.78 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |  1 |       pp16384 |       2599.01 ± 0.00 |

```
  FLASH_ATTN_EXT(hsk=64,hsv=64,nh=8,nr=4,kv=4096,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):                 23848 runs -    45.65 us/run -  33.55 MFLOP/run - 734.96 GFLOPS
  FLASH_ATTN_EXT(hsk=128,hsv=128,nh=8,nr=4,kv=4096,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):               17892 runs -    58.06 us/run -  67.11 MFLOP/run -   1.16 TFLOPS
  FLASH_ATTN_EXT(hsk=64,hsv=64,nh=8,nr=4,kv=8192,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):                 17892 runs -    56.92 us/run -  67.11 MFLOP/run -   1.18 TFLOPS
  FLASH_ATTN_EXT(hsk=128,hsv=128,nh=8,nr=4,kv=8192,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):               12682 runs -    80.56 us/run - 134.22 MFLOP/run -   1.67 TFLOPS
  FLASH_ATTN_EXT(hsk=64,hsv=64,nh=8,nr=4,kv=16384,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):                13428 runs -    78.24 us/run - 134.22 MFLOP/run -   1.72 TFLOPS
  FLASH_ATTN_EXT(hsk=128,hsv=128,nh=8,nr=4,kv=16384,nb=1,mask=1,max_bias=0.000000,logit_softcap=0.000000,prec=f32,type_KV=f16,permute=[0,1,2,3]):               8579 runs -   119.57 us/run - 268.44 MFLOP/run -   2.25 TFLOPS
  ```